ch09-authentication/requirements.txt
flask==3.0.0
bcrypt==4.1.2
PyJWT==2.8.0
